TinyTuya/requirements.txt
tinytuya>=1.13.0
paho-mqtt>=2.0.0
python-dotenv>=1.0.0